Use exclusion list for getCardsForReview and getNewCards.
AnyMemo was using largest ordinal to keep track of cached cards in the
queue so the query won't get it.
But now AnyMemo will sort the review cards by its easiness. There could be
a bug that will skip some cards for review if there is a card with large
ordinal and low easiness. The cards in between this card and old highest
ordinal will be skipped. This change will fix the problem by using
eclusion list to exclude cards in the queue. All cards in the queue will
be excluded and the query only returns cards that is not being learned.

The logic for getNewCards and getCardsForReview are similar so both are
changed for the new logic.

Add tests for the new logic.